# evcar-news/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/system/integrations": {
      "get": {
        "operationId": "AdminSystem_integrations",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/IntegrationsDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Status of every provider adapter (configured?, reason, last success/error)",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/integrations/{id}/check": {
      "post": {
        "operationId": "AdminSystem_check",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProviderCheckDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Runs a live connectivity check of one adapter (e.g. SMTP login, S3 bucket, OCM key)",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/overview": {
      "get": {
        "operationId": "AdminSystem_overview",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SystemOverviewDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin dashboard: counts, stale data, failed jobs, import jobs, unconfigured services",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/jobs": {
      "get": {
        "operationId": "AdminSystem_jobsOverview",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobsOverviewDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Background job queues (BullMQ) with counts; redis=down when Redis is unreachable",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/jobs/{queue}": {
      "get": {
        "operationId": "AdminSystem_listJobs",
        "parameters": [
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Overrides Accept-Language.",
            "schema": {
              "type": "string",
              "enum": [
                "ar",
                "en"
              ]
            }
          },
          {
            "name": "market",
            "required": false,
            "in": "query",
            "description": "Overrides X-Market.",
            "schema": {
              "example": "EG",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "default": "failed",
              "type": "string",
              "enum": [
                "waiting",
                "active",
                "completed",
                "failed",
                "delayed",
                "prioritized",
                "waiting-children"
              ]
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/JobViewDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Jobs of one queue in one state (default failed), newest first",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/jobs/{queue}/{jobId}": {
      "get": {
        "operationId": "AdminSystem_getJob",
        "parameters": [
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobDetailDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "One job with its last error (data values are not exposed)",
        "tags": [
          "admin-system"
        ]
      },
      "delete": {
        "operationId": "AdminSystem_removeJob",
        "parameters": [
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Removes a job that is not running",
        "tags": [
          "admin-system"
        ]
      }
    },
    "/api/v1/admin/system/jobs/{queue}/{jobId}/retry": {
      "post": {
        "operationId": "AdminSystem_retryJob",
        "parameters": [
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "description": "Optional client correlation id (echoed back).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobDetailDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "503": {
            "description": "Error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Retries a failed job",
        "tags": [
          "admin-system"
        ]
      }
    },
...
  },
  "info": {
    "title": "EV Car News API",
    "description": "REST API of EV Car News (evcar.news). Conventions:\n- Prefix /api/v1; admin routes under /api/v1/admin (permission-guarded), personal routes under /api/v1/me.\n- Language: ?lang=ar|en overrides Accept-Language (default ar). Market: ?market=EG overrides X-Market.\n- Single resource: {\"data\": {...}}; lists: {\"data\": [...], \"meta\": {page, pageSize, total, totalPages}}.\n- Errors: {\"error\": {\"code\", \"message\" (localized), \"details\"?, \"requestId\"}}; validation → 422 VALIDATION_FAILED.\n- Money: {\"amount\": \"decimal string\", \"currency\": \"EGP\"}; missing values are null (never 0).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 57
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "IntegrationStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "stations.open_charge_map"
          },
          "type": {
            "type": "string",
            "example": "stations",
            "description": "storage | mail | stations | availability | routing | geocoding | news | push | oauth | assistant"
          },
          "name": {
            "type": "string",
            "example": "open_charge_map"
          },
          "configured": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean",
            "description": "false when configured but inactive (e.g. console mail in production)."
          },
          "checkable": {
            "type": "boolean",
            "description": "A live check can be run with POST .../integrations/{id}/check."
          },
          "reason": {
            "type": "string",
            "description": "Why it is not configured (names missing variables, never values)."
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attribution": {
            "type": "string",
            "description": "Attribution / licence text to display."
          },
          "lastSuccessAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastError": {
            "type": "string"
          },
          "lastErrorAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "configured",
          "enabled",
          "checkable"
        ]
      },
...
      "IntegrationsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegrationStatusDto"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesDto"
          },
          "note": {
            "type": "string",
            "description": "Statuses are per API instance (last success/error are kept in memory)."
          }
        },
        "required": [
          "items",
          "capabilities",
          "note"
        ]
      },
      "ProviderCheckDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "latencyMs": {
            "type": "number"
          },
          "error": {
            "type": "string"
          },
          "skipped": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "OverviewImportsDto": {
        "type": "object",
        "properties": {
          "running": {
            "type": "number"
          },
          "failedLast7Days": {
            "type": "number"
          },
          "recent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportJobDto"
            }
          }
        },
        "required": [
          "running",
          "failedLast7Days",
          "recent"
        ]
      },
...
      "SystemOverviewDto": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "environment": {
            "type": "string"
          },
          "counts": {
            "$ref": "#/components/schemas/OverviewCountsDto"
          },
          "staleData": {
            "$ref": "#/components/schemas/StaleDataDto"
          },
          "jobs": {
            "$ref": "#/components/schemas/OverviewJobsDto"
          },
          "imports": {
            "$ref": "#/components/schemas/OverviewImportsDto"
          },
          "integrations": {
            "$ref": "#/components/schemas/OverviewIntegrationsDto"
          },
          "warnings": {
            "description": "Things an admin should look at.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "generatedAt",
          "environment",
          "counts",
          "staleData",
          "jobs",
          "imports",
          "integrations",
          "warnings"
        ]
      },
      "JobsOverviewDto": {
        "type": "object",
        "properties": {
          "redis": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ]
          },
          "workersEnabledHere": {
            "type": "boolean",
            "description": "Whether this API instance runs workers (JOBS_ENABLED)."
          },
          "queues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueSummaryDto"
            }
          }
        },
        "required": [
          "redis",
          "workersEnabledHere",
          "queues"
        ]
      },
      "JobViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "queue": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "failed"
          },
          "attemptsMade": {
            "type": "number"
          },
          "maxAttempts": {
            "type": "number"
          },
          "progress": {
            "type": "object",
            "description": "Number 0..100 or an object set by the processor."
          },
          "failedReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "processedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "delayUntil": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "queue",
          "state",
          "attemptsMade",
          "maxAttempts",
          "progress",
          "failedReason",
          "createdAt",
          "processedAt",
          "finishedAt",
          "delayUntil"
        ]
      },
      "JobDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "queue": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "example": "failed"
          },
          "attemptsMade": {
            "type": "number"
          },
          "maxAttempts": {
            "type": "number"
          },
          "progress": {
            "type": "object",
            "description": "Number 0..100 or an object set by the processor."
          },
          "failedReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "processedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "delayUntil": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "stacktrace": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dataKeys": {
            "description": "Top-level keys of the job data (values hidden).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "queue",
          "state",
          "attemptsMade",
          "maxAttempts",
          "progress",
          "failedReason",
          "createdAt",
          "processedAt",
          "finishedAt",
          "delayUntil",
          "stacktrace",
          "dataKeys"
        ]
      },
...
    }
  }
}